requirements.txt
fastapi
uvicorn
pydantic
email-validator

sqlalchemy
psycopg2-binary
alembic

python-dotenv

passlib[argon2]
argon2-cffi
python-jose[cryptography]

python-multipart
joblib

numpy
pandas
scikit-learn==1.7.2
lightgbm